Import p5-Tkx 1.09.
The Tkx module provides yet another Tk interface for Perl.  Tk is a GUI
toolkit tied to the Tcl language, and Tkx provides a bridge to Tcl that
allows Tk based applications to be written in Perl.

The main idea behind Tkx is that it is a very thin wrapper on top of
Tcl, i.e. that what you get is exactly the behaviour you read about in
the Tcl/Tk documentation with no surprises added by the Perl layer.

For a gentle introduction please read the Tkx::Tutorial.

The tkx-ed example needs the bwidget OpenBSD package.

ok sthen@