Update to talloc-2.1.2
Bite the bullet and switch to an upstream version that uses waf*.  The
plan is to use databases/tdb and devel/libtalloc in the samba4 port, and
thus to keep those ports in sync with whatever bundled version comes
with the samba4 tarball.

* trying to make sense of the waf patches is highly discouraged

While here, take maintainership.

This is a SHARED_ONLY port, and a python module is now provided.

ok sthen@