Import Texinfo 6.0
Texinfo is the official documentation format of the GNU project. It was
invented by Richard Stallman and Bob Chassell many years ago, loosely based on
Brian Reid's Scribe and other formatting languages of the time. It is used by
many non-GNU projects as well.

Texinfo uses a single source file to produce output in a number of formats,
both online and printed (dvi, html, info, pdf, xml, etc.). This means that
instead of writing different documents for online information and another for a
printed manual, you need write only one document. And when the work is revised,
you need revise only that one document.

OK kili@