Update tarsnap to 1.0.36.
Due to security and bug fixes this upgrade is strongly recommended.

This version fixes a theoretically exploitable one-byte buffer overflow
when archiving objects with long path names; a vulnerability which
allows a corrupt archive to crash tarsnap; and two bugs which cause
tarsnap to crash if the network is inoperative when it launches. This
version also adds support for performing "dry runs" without tarsnap
keys, which makes it possible to estimate tarsnap usage costs without
first creating a tarsnap account.