import p5-JSON-MaybeXS 1.003005 OK benoit@
use Cpanel::JSON::XS or JSON::XS or JSON::PP

This module first checks to see if either Cpanel::JSON::XS or
JSON::XS is already loaded, in which case it uses that module.
Otherwise it tries to load Cpanel::JSON::XS, then JSON::XS, then
JSON::PP in order, and either uses the first module it finds or
throws an error.