Even if it builds fine, it's useless at least for hs-cairo and hs-glib, because all versions of those ports that could have a change to build with ghc-7.10 and all the updated hs ports require gtk2hs-buildtools >= 0.13.11, which isn't available on hackage.haskell.org.
On the homepage (http://www.haskell.org/gtk2hs/) all you see is
'Not Found'. On the homepage mentioned in hs-cairo etc.
(http://projects.haskell.org/gtk2hs/) all you see is 'Error
establishing a database connection'.

I'm not going to pull the required parts out of github. If anyone is
interested, please investigate.

Otherwise, this one and everything depending on it will be removed
in a week.