import pdfpc-4.0.0
pdfpc is a GTK-based presentation viewer which uses Keynote-like
multi-monitor output to provide meta information to the speaker during
the presentation. It is able to show a normal presentation window on one
screen while showing a more sophisticated overview on the other one,
providing information like a picture of the next slide, as well as the
time left in the presentation. pdfpc processes PDF documents, which can
be created using nearly all modern presentation software.

ok aja@