Import wxglade-0.7.0, from maintainer Daniel Winters.
Feedback from landry@ and sthen@; ok sthen@

pkg/DESCR:

wxGlade is a GUI designer written in Python with the popular GUI
toolkit wxPython, that helps you create wxWidgets/wxPython user
interfaces. At the moment it can generate Python, C++, Perl, Lisp and
XRC (wxWidgets' XML resources) code.

As you can guess by the name, its model is Glade, the famous
GTK+/GNOME GUI builder, with which wxGlade shares the philosophy and
the look & feel (but not a line of code).

It is not (and will never be) a full featured IDE, but simply a
"designer": the generated code does nothing apart from displaying the
created widgets.