Adjust the icinga2 port:
- Replace previous backout with fixes from Gunnar Beutner (upstream):

 . Make sure the timer thread is gone when calling fork()
 . Ensure that the main thread pool gets re-initialized properly after fork()

- Add a check so that 'make test' fails early with a useful error if the
libconfig package is installed; the objects in the build directory don't
have a finalized rpath and they are looking in /usr/local/lib before the
build directory, so the wrong libconfig so is picked up.

I think this port should now be in good shape, please let me know if you're
using it and run into problems with forking etc.