Stop using RWX memory for closures.
This switches libffi to the use of two different mappings per closure,
one with RW access, the other with RX access.  Even if not perfect it is
already an improvement.  Soon RWX memory will be unavailable by default,
so fixing libffi means that we don't have to tweak an unknown percentage
of the 565 ports that link against it...

Initial diff from sthen@ who also did most of the tests.
ok sthen@ jasper@ (maintainer)