Import GpgME++, an optional but useful dependency for KF5Wallet.
GpgME++ is a C++ wrapper (or C++ bindings) for the GnuPG project's gpgme (GnuPG
Made Easy) library, version 0.4.4 and later.

It is fairly complete, with some minor things still missing (in particular, the
key edit interface). It is mostly tested using external event loops, for which
its design is optimised.

As a part of KF5 import: tested by kirby@ and awolk@, final okay awolk@.