Import U-Boot 2016.05
U-Boot is a firmware for embedded boards based on PowerPC, ARM, MIPS and
several other processors, which can be installed in a boot ROM and used to
initialize and test the hardware or to download and run application code.

As ARM systems often do not come with firmware it must be supplied on
an SD card or MMC device to have a bootable system.

This port provides U-Boot for various boards using ARM processors.

armv7 requires the support for EFI payloads added in this release to run
the bootloader.  A proposed patch to prevent the EFI interface from
allocating pages from unpopulated memory by Alexander Graf is included
which will hopefully be part of future releases:
http://marc.info/?l=u-boot&m=146434472023891&w=2

Feedback from jca@ and sthen@.  ok sthen@ on an earlier version without
the EFI patch.