Unbreak, from semarie:
When using --enable-local-rebuild option, the configure script will
check the existence of rustc binary from the local-rust-root directory
(/usr/local by default). It is the binary the build process will copy in
stage0 directory.

With the diff, we explicity use the bootstrap directory for configuring.
Note that we still we manually copying all the bootstrap files in
post-configure, as the standard build system will *not* copy the
libraries in lib/ which could be required to run rustc binary
(libraries like libc.so used by bootstrapper).