Import net/iperf3, a rewritten version of iperf.
iperf is a tool for active measurements of the maximum achievable bandwidth on
IP networks.  It supports tuning of various parameters related to timing,
protocols, and buffers.  For each test it reports the bandwidth, loss, and other
parameters.

This version, sometimes referred to as iperf3, is a redesign of an original
version developed at NLANR / DAST.  iperf3 is a new implementation from scratch,
with the goal of a smaller, simpler code base, and a library version of the
functionality that can be used in other programs. iperf3 also incorporates a
number of features found in other tools such as nuttcp and netperf, but were
missing from the original iperf.  Note that iperf3 is not backwards compatible
with the original iperf.

prodding & patches from jca@, thanks!
feedback & ok jca@ sthen@