Import libnspsl-0.1.0.
The Public Suffix List is a database of top level domain names. The
database allows an application to determine if if a domain name requires
an additional label to be valid.

The nspsl library generates a static code representation of the Public
Suffix List for use in the NetSurf web browser.

ok landry@