Import qtbase-docs, the qtbase documentation.
It's based on the same distfile as qtbase package, but uses totally
different build mechanizm. And it's unreal to build -qch part without
circular dependency with qttools anyway.

Do not try to build this one until it's hooked to the build: this port
relies on functionality to be committed (really soon).