add pledges for dig/host/nslookup in the ports version of BIND. initial pledge is "stdio rpath inet unix dns", dropping to "stdio inet dns" after argument parsing.
access to resolv.conf is required late; the dns pledge is used for this
rather than requiring full rpath; however contrary to the version in
base, inet is allowed as well, so that it can be used as a debug tool
for servers on alternate ports.

works fine for me; no feedback after posting yet so committing to get
real-world testing. please report any issues.