import botan2 2.0.1 input and OK jca@
Comment:
crypto and TLS for C++11

Description:
Botan (Japanese for peony) is a cryptography library written in
C++11 and released under the permissive Simplified BSD license.
Botan's goal is to be the best option for cryptography in new C++
code by offering the tools necessary to implement a range of practical
systems, such as TLS/DTLS, PKIX certificate handling, PKCS#11 and
TPM hardware support, password hashing, and post quantum crypto
schemes.  In addition to the C++, botan has a C89 API specifically
designed to be easy to call from other languages.  A Python binding
using ctypes calling the C89 API is included.