Remove www/elinks from the ports tree.
The last release of elinks was in 2009, and it has since come to our attention
that it does not verify the authenticity of SSL certificates:
https://marc.info/?l=elinks-dev&m=148896031830582&w=2

We've not heard from upstream, so for now we are removing elinks from ports.

For a more modern text-based browser with SSL certificate verification, see
www/lynx.

OK jca@