Delete a comment about USE_GROFF; no package change, no bump.
Formatting with groff and mandoc is now almost identical, and given
that this is DocBook stuff, there is no reason to assume that it
will break more badly than similar ports in the future.

Mandoc ignores the .sp requests in the table and issues UNSUPP
messages about them, but actually, they are pointless and the table
looks better without them.