Import converters/p5-Date-Tolkien-Shire-Data ok benoit@
This Perl module contains common or near-common code factored out of
Date::Tolkien::Shire and DateTime::Fiction::JRRTolkien::Shire. This is
mostly data and calculations common to the two modules, but there is
also a piece of new functionality: an output formatter for Shire dates
analogous to strftime(). All functionality is provided by subroutines.

Because of its nature, you probably do not want to use this module
directly.  At least, not without looking into Date::Tolkien::Shire which
is a semi-thin object-oriented front end to this module, or
DateTime::Fiction::JRRTolkien::Shire which is a somewhat-quirky member
of the DateTime family.