Import x11/xcape:
---8<---
xcape allows you to use a modifier key as another key when pressed and released
on its own. Note that it is slightly slower than pressing the original key,
because the pressed event does not occur until the key is released. The default
behaviour is to generate the Escape key when Left Control is pressed and
released on its own. (If you don't understand why anybody would want this, I'm
guessing that Vim is not your favourite text editor ;)
--->8---

From Jon Bernard, with input from myself, sthen@ and schwarze@.

OK sthen@.

Thanks all.