Import multimedia/streamlink, a program that lets you pipe video streams into a video player. with and ok nigel@
Streamlink is a command-line utility that pipes video streams from      
various services into a video player, such as VLC. The main purpose of  
Streamlink is to allow the user to avoid buggy and CPU heavy flash      
plugins but still be able to enjoy various streamed content. There is   
also an API available for developers who want access to the video stream                                                                         
data. This project was forked from Livestreamer, which is no longer     
maintained.                         

Streamlink is built upon a plugin system which allows support for new   
services to be easily added. Currently most of the big streaming        
services are supported, such as:    

* Dailymotion                       
* Livestream                        
* Twitch                            
* UStream                           
* YouTube Live                      

and many more.