import p5-Forest 0.10 from Matthias Pitzl; OK sthen@
Comment:
a collection of n-ary tree related modules for Perl

Description:
Forest is intended to be a replacement for the Tree::Simple family
of modules, and fixes many of the issues that have always bothered
me about them.  It is by no means a complete replacement yet, but
should eventually grow to become that.