Import koboredux-0.7.4.
Kobo Redux is a frantic '80s style 2D shooter with the look and feel a
'90s arcade cabinet, the smoothness of current technology, and a matching
soundtrack. The gameplay is fast, intense and unforgiving, while trying to
tone down the frustrating quirkiness of the actual games of the '80s. A true
challenge in the spirit of the arcade era!

The full game, which provides improved graphics and music, can be purchased
from: https://olofson.itch.io/kobo-redux

tweaks + ok bcallah@