Import wireless, from Gregor Best (upstream) tweaks from czarkoff and sthen, ok sthen a while ago
From DESCR:
wireless scans for and configures wireless networks on OpenBSD.  It uses
ifconfig(8) to do most of the heavy lifting and basically just
interprets and applies scan results.

and to quote sthen:
I know there has been a desire to have something in base to handle this
type of configuration for a while, and that would still be nice to have,
but as-is this is quite an improvement over the shell scripts that I'm
sure many of us have written (especially in terms of handling networks
with dubious SSIDs).