Fix a crash in libnetsnmp client, occurring in some circumstances where a local IP address is set. Cause is a use-after-free, OpenBSD's malloc helpfully often picks these up early (with relatively low impact on performance), often turning what would be a hidden bug into a likely crash.
Reported and analysis from Andrea Biscuola of Zabbix,
https://support.zabbix.com/browse/ZBX-12726
https://sourceforge.net/p/net-snmp/bugs/2803/

Diff from me. I was going to wait for confirmation from Andrea that it
fixes the issue, but having read more of the code I'm confident that this
change or something like it is needed, so committing now to make it pre 6.2.
ok naddy