Switch lang/go-bootstrap to use the go1.4 release tarball.
Upstream is providing a tarball of the release-branch.go1.4 git branch,
which contains minor fixes to keep Go 1.4 building on current platforms.
Most of the patches that we currently have for lang/go-bootstrap have
now been committed upstream.

The current go1.4 release tarball contains one of the patches and most of
the remaining patches will be removed when a new release tarball becomes
available.

ok sthen@