Import mp3applygain 1.1.0
This fills a void left since the removal of mp3gain.  It applies replaygain
information to mp3 audio data, which allows for the use of replaygain
information even on players that do not support it via tags (common in
hardware mp3 players).  mp3applygain is a fork of mp3gain with all code
removed except the code to actually modify the mp3 audio data to apply
the gain.

Help from jca@
OK sthen@