Switch to using estraier gem
This is much simpler than trying to build the support that ships with
the hyperestaier distfile.  No significant difference in the ruby code,
even though it appears to be 3 tiny versions behind (1.4.10 vs 1.4.13).

This should make it easier to support future ruby versions, as previously
those required manual changes to the port Makefile.