Import kirigami2
"New framework: kirigami, a set of QtQuick plugins to build user
interfaces based on the KDE UX guidelines"

https://www.kde.org/announcements/kde-frameworks-5.37.0.php

$ pkg/DESCR
Kirigami is a set of QtQuick components at the moment targeted for mobile use
targeting both Plasma Mobile and Android. It’s not a whole set of components,
all the “Primitive” ones like buttons and textboxes are a job for
QtQuickControls but it’s a set of high level components to make the creation of
applications that look and feel great on mobile as well as desktop devices and
follow the Kirigami Human Interface Guidelines.

ok landry@