Import fifechan-0.1.4.
Fifechan is a lightweight cross platform GUI library written in C++
specifically designed for games. It has a small yet powerful built in set of
extendable GUI Widgets allowing users to create virtually unlimited types of
widgets. Fifechan supports rendering in SDL, OpenGL, or Allegro out of the
box or it can be adapted to use any rendering engine the user requires.
Events are pushed to Fifechan which allows users to use any input library
they wish or they could use the built in input handling through either SDL
input or Allegro input. The primary goal for Fifechan is to keep it
extendable, lightweight and still be powerful enough to use in all types of
games out of the box.

ok kirby@