Import deutex-5.1.0.
DeuTex is a WAD file composer. It can do many things with Doom, Freedoom,
Heretic, Hexen, and Strife WAD files, such as extracting and inserting
graphics, sounds, levels, and other resources. It can be used for creating
and modifying IWAD ("Internal WAD") and PWAD ("Patch WAD") files both.

It is a command-line driven program and as such, is suitable for scripting
such as from a shell script or Makefile.

ok landry@