Append Linux x86_64 to the User-Agent:
Mozilla/5.0 (X11; OpenBSD amd64; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/62.0.3202.94 Safari/537.36

This is required because a lot of big service vendors are misusing the User-Agent
to discriminate non-mainstream OS users. A good example is Google and Microsoft
not allowing non-mainstream OS users to use features that are NOT OS dependant,
like Google Maps full-mode or Skype web.

All of the browsers have a proper way of checking what they support so relying
on the User-Agent string is completely stupid because what works on Linux, will
99% work on other operating systems as well, since the browser supports it.
(e.g. 3d acceleration)

Instead of completely saying that we are Linux, it is enough to append Linux to
the User-Agent platform string, so people relying on the User-Agent for filtering,
statistics can still filter for OpenBSD.

Everyone can thank these service vendors for fiddling with the User-Agent in a way
that they weren't supposed to.

discussed with deraadt@