Repair arch tests and maybe unlock neomutt on powerpc
On powerpc, MACHINE_ARCH is "powerpc" and ARCH is "socppc" or "macppc",
the test fails and neomutt tries to use kyotocabinet, marked as BROKEN.
Fix test and use qdbm as intented.  On arm, where MACHINE_ARCH is "arm" and
ARCH is "armv7", the test fails thus neomutt is already using kyotocabinet
(which builds on arm now), so remove the test.  No need for a bump since
the build failed in both cases.

Reported by "gumnos", ok sthen@ (maintainer)