Import games/returntotheroots, an open source engine remake of The Settlers 2. ok awolk@
Return To The Roots is a fan project which aims to renew the original
The Settlers 2.

We aim to create new features such as a multiplayer mode via the
Internet as well as support for modern hardware.

You will need an original copy of "The Settlers 2 Gold Edition" to play
Return To The Roots.

The Settlers 2 is a 4X game similar to Civilization and Alpha Centauri.

Comes with a game client (s25client) and map editor (s25edit).