Import libtorrent-rasterbar-1.1.6.
libtorrent(-rasterbar) is a C++ library that aims to be a good
alternative to all the other bittorrent implementations around.

The main goals of libtorrent are to be cpu efficient, memory efficient
and very easy to use.

Note that this is NOT the same library as the libtorrent package!


Based on a port by dcoppa@
Tweaks + ok bcallah@